+++ KNN/src/structure/Edge.java
@@ -2,14 +2,14 @@ package structure;
 
 public class Edge{
 	Vertex v;
-	int w;
+	double w;
 	
 	Edge(Vertex v){
 		this.v = v;
 		this.w = 0;
 	}
 	
-	Edge(Vertex v,int w){
+	Edge(Vertex v,double w){
 		this.v = v;
 		this.w = w;
 	}
